feat(UsersShowcase): New component created
